Move addSystemViewColumns from ViewInfo to QueryModel/utils
